arduinoTest/requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
pyserial
gevent
plotly
